ISSUE-82 Address ingestion issue(s) with eTUFF-tuna590051-debug.txt
