+++ .github/workflows/build.yml
@@ -77,6 +77,8 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python }}
+      - name: Upgrade pip
+        run: pip install --upgrade pip
       - name: Install tox and any other packages
         run: pip install tox
       - name: Run tox
