+++ tagbase_server/tox.ini
@@ -12,6 +12,7 @@ python =
     3.10: py310
 
 [testenv]
+pip = pip install --upgrade pip
 deps = -r test-requirements.txt
 commands = pytest --cov=tagbase_server --cov-report xml:tagbase_server/coverage.xml --cov-config=tox.ini --cov-branch --cov-report=term-missing
